+++ accounts/urls.py
@@ -4,8 +4,8 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.login, name='login'),
+    #path('', views.login, name='login'),
     path('', views.register, name='register'),
-    path('', views.logout, name='logout'),
+    #path('', views.logout, name='logout'),
     
 ]
