+++ login/urls.py
@@ -1,0 +1,11 @@
+
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('', views.login, name='login'),
+
+    path('logout', views.logout, name='logout'),
+    
+]
