+++ templates/login.html
@@ -250,11 +250,11 @@
             <div class="container">
                 <div class="row">
                     <div class="col-md-6 copyright">
-                        <p>Copyright &copy; <a href="https://htmlcodex.com">HTML Codex</a>. All Rights Reserved</p>
+                        <p>Copyright &copy; <a href="https://htmlcodex.com">Jahangir</a>. All Rights Reserved</p>
                     </div>
 
                     <div class="col-md-6 template-by">
-                        <p>Template By <a href="https://htmlcodex.com">HTML Codex</a></p>
+                        <p>Template By <a href="https://htmlcodex.com">Jahangir</a></p>
                     </div>
                 </div>
             </div>
